Fix navigation into lists, seqs, and sets
In this commit, the Browsable protocol is moved to a separate namespace, so that
it can be used by gadget.datafy/nav-in.

The Browsable `entries` function is responsible for generating pairs of keys and
data which can be clicked in the UI for further navigation. With this change,
nav-in now converts the datastructure to its browsable entries before looking up
the provided key.

This fixes navigation into lists, and seqs, which was previously not working
because nav-in tried to `get` from the raw data directly. It also fixes
navigation for sets by implementing the Browsable protocol for them, and makes
sure any user space/future extensions based on Browsable will, in fact, be
browsable.